Split previous core.py into two modules user/task.py and user/users.py
