+++ locust/test/test_taskratio.py
@@ -1,6 +1,6 @@
 import unittest
 
-from locust.core import User, TaskSet, task
+from locust.user import User, TaskSet, task
 from locust.inspectlocust import get_task_ratio_dict
