+++ locust/user/__init__.py
@@ -1,0 +1,2 @@
+from .task import task, TaskSet
+from .users import HttpUser, User
